2021/src/test/java/de/jfriemel/aoc/days/Day05Test.java
package de.jfriemel.aoc.days;

import org.junit.Test;

import java.util.Arrays;
import java.util.List;

import static org.junit.Assert.assertEquals;

public class Day05Test {

    private static final String INPUT_STRING = """
            0,9 -> 5,9
            8,0 -> 0,8
            9,4 -> 3,4
            2,2 -> 2,1
            7,0 -> 7,4
            6,4 -> 2,0
            0,9 -> 2,9
            3,4 -> 1,4
            0,0 -> 8,8
            5,5 -> 8,2""";

    private static final List<String> INPUT = Arrays.asList(INPUT_STRING.split("\n"));

    @Test
    public void testPart1() {
        String result = new Day05().part1(INPUT);
        assertEquals("5", result);
    }

    @Test
    public void testPart2() {
        String result = new Day05().part2(INPUT);
        assertEquals("12", result);
    }

}
